+++ lib/result_dart.dart
@@ -1,7 +1,6 @@
 library result_dart;
 
 export 'src/async_result.dart';
-export 'src/function.dart';
 export 'src/result.dart';
 export 'src/result_extension.dart';
 export 'src/unit.dart';