+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## [1.0.2] - 2021-12-18
+
+* fix: separed functions.dart import
+
+## [1.0.1] - 2021-12-17
+
+* fix: recover operator return a `Result` instead a `Failure`.
+
 ## [1.0.0+2] - 2021-12-16
 
 * Initial release
-## [1.0.1] - 2021-12-17
 
-* fix: recover operator return a `Result` instead a `Result`.